+++ versions.tf
@@ -1,6 +1,6 @@
 
 terraform {
-  required_version = ">= 1.0.0"
+  required_version = ">= 1.3.0"
 
   required_providers {
     aci = {
@@ -8,6 +8,4 @@ terraform {
       version = ">= 2.0.0"
     }
   }
-
-  experiments = [module_variable_optional_attrs]
 }